added handle update customer order

+++ src/pages/CartDetails/CartDetails.jsx
@@ -3,6 +3,8 @@ import carDetailBgImg from "../../assets/images/cart_details/cart-details.svg";
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../providers/AuthProvider";
 import OrderItem from "../../components/unique/OrderItem/OrderItem";
+import swal from "sweetalert";
+
 const CartDetails = () => {
   const [orderItems, setOrderItems] = useState([]);
   const { user } = useContext(AuthContext);
@@ -18,14 +20,43 @@ const CartDetails = () => {
   }, [url]);
 
   const handleDeleteOrderItem = (id) => {
-    fetch(`http://localhost:5000/items/${id}`, {
-      method: "DELETE",
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this imaginary file!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        fetch(`http://localhost:5000/items/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.acknowledged) {
+              const reminig = orderItems.filter((item) => item._id !== id);
+              setOrderItems(reminig);
+            }
+          });
+      }
+    });
+  };
+
+  const handleUpdateOrder = (id) => {
+    fetch(`http://localhost:5000/orders/${id}`, {
+      method: "PATCH",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ status: "Confirmed" }),
     })
       .then((res) => res.json())
       .then((data) => {
-        if(data.acknowledged){
-            const reminig = orderItems.filter(item=> item._id !== id)
-            setOrderItems(reminig)
+        console.log(data);
+        if (data.acknowledged) {
+          const reminigItems = orderItems.filter((item) => item._id !== id);
+          const updatedItem = orderItems.find((item) => item._id === id);
+          updatedItem.status = "Confirmed";
+          const newItems = [updatedItem, ...reminigItems];
+          setOrderItems(newItems);
         }
       });
   };
@@ -41,6 +72,7 @@ const CartDetails = () => {
                 <OrderItem
                   key={item._id}
                   handleDeleteOrderItem={handleDeleteOrderItem}
+                  handleUpdateOrder={handleUpdateOrder}
                   item={item}
                 />
               ))}
